refacto: Improve project structure
