+++ lib/slack.py
@@ -2,7 +2,6 @@ import os
 import hmac
 import hashlib
 from datetime import datetime
-from .database import get_db_connection
 
 def verify_slack_request(timestamp, body, signature):
     """Verify that the request actually came from Slack"""
@@ -18,15 +17,3 @@ def verify_slack_request(timestamp, body, signature):
 
     return hmac.compare_digest(my_signature, signature)
 
-def is_admin(user_id):
-    """Check if a user is an admin"""
-    conn = get_db_connection()
-    cur = conn.cursor()
-    
-    cur.execute('SELECT EXISTS(SELECT 1 FROM admin_users WHERE user_id = %s)', (user_id,))
-    is_admin = cur.fetchone()[0]
-    
-    cur.close()
-    conn.close()
-    
-    return is_admin
